sq-be/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
requests==2.31.0
pytest==7.4.3

